requirements.txt
streamlit>=1.32
torch>=2.2
torchvision>=0.17
numpy
pillow
matplotlib
seaborn
scikit-learn
opencv-python-headless
tqdm
